update project model to use embedded data fields
